// src/generated/grpc/apis/massa/api/v1/public_pb.d.ts
  export type AsObject = {
    call?: massa_model_v1_execution_pb.ReadOnlyExecutionCall.AsObject,
  }
// ...
  export enum FilterCase { 
    FILTER_NOT_SET = 0,
    ADDRESS_KEY = 1,
  }
// ...
export class NewSlotExecutionOutputsFilter extends jspb.Message {
  getStatus(): massa_model_v1_execution_pb.ExecutionOutputStatus;
  setStatus(value: massa_model_v1_execution_pb.ExecutionOutputStatus): NewSlotExecutionOutputsFilter;

  getSlotRange(): massa_model_v1_slot_pb.SlotRange | undefined;
  setSlotRange(value?: massa_model_v1_slot_pb.SlotRange): NewSlotExecutionOutputsFilter;
  hasSlotRange(): boolean;
  clearSlotRange(): NewSlotExecutionOutputsFilter;

  getAsyncPoolChangesFilter(): AsyncPoolChangesFilter | undefined;
  setAsyncPoolChangesFilter(value?: AsyncPoolChangesFilter): NewSlotExecutionOutputsFilter;
  hasAsyncPoolChangesFilter(): boolean;
  clearAsyncPoolChangesFilter(): NewSlotExecutionOutputsFilter;

  getExecutedDenounciationFilter(): ExecutedDenounciationFilter | undefined;
  setExecutedDenounciationFilter(value?: ExecutedDenounciationFilter): NewSlotExecutionOutputsFilter;
  hasExecutedDenounciationFilter(): boolean;
  clearExecutedDenounciationFilter(): NewSlotExecutionOutputsFilter;

  getEventFilter(): ExecutionEventFilter | undefined;
  setEventFilter(value?: ExecutionEventFilter): NewSlotExecutionOutputsFilter;
  hasEventFilter(): boolean;
  clearEventFilter(): NewSlotExecutionOutputsFilter;

  getExecutedOpsChangesFilter(): ExecutedOpsChangesFilter | undefined;
  setExecutedOpsChangesFilter(value?: ExecutedOpsChangesFilter): NewSlotExecutionOutputsFilter;
  hasExecutedOpsChangesFilter(): boolean;
  clearExecutedOpsChangesFilter(): NewSlotExecutionOutputsFilter;

  getLedgerChangesFilter(): LedgerChangesFilter | undefined;
  setLedgerChangesFilter(value?: LedgerChangesFilter): NewSlotExecutionOutputsFilter;
  hasLedgerChangesFilter(): boolean;
  clearLedgerChangesFilter(): NewSlotExecutionOutputsFilter;

  getFilterCase(): NewSlotExecutionOutputsFilter.FilterCase;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): NewSlotExecutionOutputsFilter.AsObject;
  static toObject(includeInstance: boolean, msg: NewSlotExecutionOutputsFilter): NewSlotExecutionOutputsFilter.AsObject;
  static serializeBinaryToWriter(message: NewSlotExecutionOutputsFilter, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): NewSlotExecutionOutputsFilter;
  static deserializeBinaryFromReader(message: NewSlotExecutionOutputsFilter, reader: jspb.BinaryReader): NewSlotExecutionOutputsFilter;
}

export namespace NewSlotExecutionOutputsFilter {
  export type AsObject = {
    status: massa_model_v1_execution_pb.ExecutionOutputStatus,
    slotRange?: massa_model_v1_slot_pb.SlotRange.AsObject,
    asyncPoolChangesFilter?: AsyncPoolChangesFilter.AsObject,
    executedDenounciationFilter?: ExecutedDenounciationFilter.AsObject,
    eventFilter?: ExecutionEventFilter.AsObject,
    executedOpsChangesFilter?: ExecutedOpsChangesFilter.AsObject,
    ledgerChangesFilter?: LedgerChangesFilter.AsObject,
  }

  export enum FilterCase { 
    FILTER_NOT_SET = 0,
    STATUS = 1,
    SLOT_RANGE = 2,
    ASYNC_POOL_CHANGES_FILTER = 3,
    EXECUTED_DENOUNCIATION_FILTER = 4,
    EVENT_FILTER = 5,
    EXECUTED_OPS_CHANGES_FILTER = 6,
    LEDGER_CHANGES_FILTER = 7,
  }
}

export class AsyncPoolChangesFilter extends jspb.Message {
  getNone(): massa_model_v1_commons_pb.Empty | undefined;
  setNone(value?: massa_model_v1_commons_pb.Empty): AsyncPoolChangesFilter;
  hasNone(): boolean;
  clearNone(): AsyncPoolChangesFilter;

  getType(): massa_model_v1_execution_pb.AsyncPoolChangeType;
  setType(value: massa_model_v1_execution_pb.AsyncPoolChangeType): AsyncPoolChangesFilter;

  getHandler(): string;
  setHandler(value: string): AsyncPoolChangesFilter;

  getDestinationAddress(): string;
  setDestinationAddress(value: string): AsyncPoolChangesFilter;

  getEmitterAddress(): string;
  setEmitterAddress(value: string): AsyncPoolChangesFilter;

  getCanBeExecuted(): boolean;
  setCanBeExecuted(value: boolean): AsyncPoolChangesFilter;

  getFilterCase(): AsyncPoolChangesFilter.FilterCase;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): AsyncPoolChangesFilter.AsObject;
  static toObject(includeInstance: boolean, msg: AsyncPoolChangesFilter): AsyncPoolChangesFilter.AsObject;
  static serializeBinaryToWriter(message: AsyncPoolChangesFilter, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): AsyncPoolChangesFilter;
  static deserializeBinaryFromReader(message: AsyncPoolChangesFilter, reader: jspb.BinaryReader): AsyncPoolChangesFilter;
}

export namespace AsyncPoolChangesFilter {
  export type AsObject = {
    none?: massa_model_v1_commons_pb.Empty.AsObject,
    type: massa_model_v1_execution_pb.AsyncPoolChangeType,
    handler: string,
    destinationAddress: string,
    emitterAddress: string,
    canBeExecuted: boolean,
  }

  export enum FilterCase { 
    FILTER_NOT_SET = 0,
    NONE = 1,
    TYPE = 2,
    HANDLER = 3,
    DESTINATION_ADDRESS = 4,
    EMITTER_ADDRESS = 5,
    CAN_BE_EXECUTED = 6,
  }
}
// ...
export class ExecutionEventFilter extends jspb.Message {
  getNone(): massa_model_v1_commons_pb.Empty | undefined;
  setNone(value?: massa_model_v1_commons_pb.Empty): ExecutionEventFilter;
  hasNone(): boolean;
  clearNone(): ExecutionEventFilter;

  getCallerAddress(): string;
  setCallerAddress(value: string): ExecutionEventFilter;

  getEmitterAddress(): string;
  setEmitterAddress(value: string): ExecutionEventFilter;

  getOriginalOperationId(): string;
  setOriginalOperationId(value: string): ExecutionEventFilter;

  getIsFailure(): boolean;
  setIsFailure(value: boolean): ExecutionEventFilter;

  getFilterCase(): ExecutionEventFilter.FilterCase;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): ExecutionEventFilter.AsObject;
  static toObject(includeInstance: boolean, msg: ExecutionEventFilter): ExecutionEventFilter.AsObject;
  static serializeBinaryToWriter(message: ExecutionEventFilter, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): ExecutionEventFilter;
  static deserializeBinaryFromReader(message: ExecutionEventFilter, reader: jspb.BinaryReader): ExecutionEventFilter;
}

export namespace ExecutionEventFilter {
  export type AsObject = {
    none?: massa_model_v1_commons_pb.Empty.AsObject,
    callerAddress: string,
    emitterAddress: string,
    originalOperationId: string,
    isFailure: boolean,
  }

  export enum FilterCase { 
    FILTER_NOT_SET = 0,
    NONE = 1,
    CALLER_ADDRESS = 2,
    EMITTER_ADDRESS = 3,
    ORIGINAL_OPERATION_ID = 4,
    IS_FAILURE = 5,
  }
}

export class ExecutedOpsChangesFilter extends jspb.Message {
  getNone(): massa_model_v1_commons_pb.Empty | undefined;
  setNone(value?: massa_model_v1_commons_pb.Empty): ExecutedOpsChangesFilter;
  hasNone(): boolean;
  clearNone(): ExecutedOpsChangesFilter;

  getOperationId(): string;
  setOperationId(value: string): ExecutedOpsChangesFilter;

  getFilterCase(): ExecutedOpsChangesFilter.FilterCase;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): ExecutedOpsChangesFilter.AsObject;
  static toObject(includeInstance: boolean, msg: ExecutedOpsChangesFilter): ExecutedOpsChangesFilter.AsObject;
  static serializeBinaryToWriter(message: ExecutedOpsChangesFilter, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): ExecutedOpsChangesFilter;
  static deserializeBinaryFromReader(message: ExecutedOpsChangesFilter, reader: jspb.BinaryReader): ExecutedOpsChangesFilter;
}

export namespace ExecutedOpsChangesFilter {
  export type AsObject = {
    none?: massa_model_v1_commons_pb.Empty.AsObject,
    operationId: string,
  }

  export enum FilterCase { 
    FILTER_NOT_SET = 0,
    NONE = 1,
    OPERATION_ID = 2,
  }
}

export class ExecutedDenounciationFilter extends jspb.Message {
  getNone(): massa_model_v1_commons_pb.Empty | undefined;
  setNone(value?: massa_model_v1_commons_pb.Empty): ExecutedDenounciationFilter;
  hasNone(): boolean;
  clearNone(): ExecutedDenounciationFilter;

  getFilterCase(): ExecutedDenounciationFilter.FilterCase;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): ExecutedDenounciationFilter.AsObject;
  static toObject(includeInstance: boolean, msg: ExecutedDenounciationFilter): ExecutedDenounciationFilter.AsObject;
  static serializeBinaryToWriter(message: ExecutedDenounciationFilter, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): ExecutedDenounciationFilter;
  static deserializeBinaryFromReader(message: ExecutedDenounciationFilter, reader: jspb.BinaryReader): ExecutedDenounciationFilter;
}

export namespace ExecutedDenounciationFilter {
  export type AsObject = {
    none?: massa_model_v1_commons_pb.Empty.AsObject,
  }

  export enum FilterCase { 
    FILTER_NOT_SET = 0,
    NONE = 1,
  }
}

export class LedgerChangesFilter extends jspb.Message {
  getNone(): massa_model_v1_commons_pb.Empty | undefined;
  setNone(value?: massa_model_v1_commons_pb.Empty): LedgerChangesFilter;
  hasNone(): boolean;
  clearNone(): LedgerChangesFilter;

  getAddress(): string;
  setAddress(value: string): LedgerChangesFilter;

  getFilterCase(): LedgerChangesFilter.FilterCase;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): LedgerChangesFilter.AsObject;
  static toObject(includeInstance: boolean, msg: LedgerChangesFilter): LedgerChangesFilter.AsObject;
  static serializeBinaryToWriter(message: LedgerChangesFilter, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): LedgerChangesFilter;
  static deserializeBinaryFromReader(message: LedgerChangesFilter, reader: jspb.BinaryReader): LedgerChangesFilter;
}

export namespace LedgerChangesFilter {
  export type AsObject = {
    none?: massa_model_v1_commons_pb.Empty.AsObject,
    address: string,
  }

  export enum FilterCase { 
    FILTER_NOT_SET = 0,
    NONE = 1,
    ADDRESS = 2,
  }
}